Revert "fix(emulation): sync navigator.platform with custom user agent across…"
This reverts commit 0f88339230800d896e100c29d13e8aa4352ba1b3.
