+++ packages/playwright-core/src/server/browserContext.ts
@@ -790,74 +790,6 @@ export function normalizeProxySettings(proxy: types.ProxySettings): types.ProxyS
   return { ...proxy, server, bypass };
 }
 
-// Chromium reference: https://source.chromium.org/chromium/chromium/src/+/main:components/embedder_support/user_agent_utils.cc;l=434
-export function calculateUserAgentEmulation(options: types.BrowserContextOptions): {
-  navigatorPlatform: string | undefined;
-  userAgentMetadata: {
-    mobile: boolean;
-    model: string;
-    architecture: string;
-    platform: string;
-    platformVersion: string;
-  } | undefined;
-} {
-  const ua = options.userAgent;
-  if (!ua)
-    return { navigatorPlatform: undefined, userAgentMetadata: undefined };
-
-  const userAgentMetadata = {
-    mobile: !!options.isMobile,
-    model: '',
-    architecture: 'x86',
-    platform: 'Windows',
-    platformVersion: '',
-  };
-
-  const androidMatch = ua.match(/Android (\d+(\.\d+)?(\.\d+)?)/);
-  const iPhoneMatch = ua.match(/iPhone OS (\d+(_\d+)?)/);
-  const iPadMatch = ua.match(/iPad; CPU OS (\d+(_\d+)?)/);
-  const macOSMatch = ua.match(/Mac OS X (\d+(_\d+)?(_\d+)?)/);
-  const windowsMatch = ua.match(/Windows\D+(\d+(\.\d+)?(\.\d+)?)/);
-  if (androidMatch) {
-    userAgentMetadata.platform = 'Android';
-    userAgentMetadata.platformVersion = androidMatch[1];
-    userAgentMetadata.architecture = 'arm';
-  } else if (iPhoneMatch) {
-    userAgentMetadata.platform = 'iOS';
-    userAgentMetadata.platformVersion = iPhoneMatch[1].replace(/_/g, '.');
-    userAgentMetadata.architecture = 'arm';
-  } else if (iPadMatch) {
-    userAgentMetadata.platform = 'iOS';
-    userAgentMetadata.platformVersion = iPadMatch[1].replace(/_/g, '.');
-    userAgentMetadata.architecture = 'arm';
-  } else if (macOSMatch) {
-    userAgentMetadata.platform = 'macOS';
-    userAgentMetadata.platformVersion = macOSMatch[1].replace(/_/g, '.');
-    if (!ua.includes('Intel'))
-      userAgentMetadata.architecture = 'arm';
-  } else if (windowsMatch) {
-    userAgentMetadata.platform = 'Windows';
-    userAgentMetadata.platformVersion = windowsMatch[1];
-  } else if (ua.toLowerCase().includes('linux')) {
-    userAgentMetadata.platform = 'Linux';
-  }
-  if (ua.includes('ARM') || ua.includes('aarch64'))
-    userAgentMetadata.architecture = 'arm';
-
-  let navigatorPlatform: string | undefined;
-  if (!process.env.PLAYWRIGHT_NO_UA_PLATFORM) {
-    switch (userAgentMetadata.platform) {
-      case 'Android': navigatorPlatform = userAgentMetadata.architecture === 'arm' ? 'Linux armv8l' : 'Linux x86_64'; break;
-      case 'iOS': navigatorPlatform = ua.includes('iPad') ? 'iPad' : 'iPhone'; break;
-      case 'macOS': navigatorPlatform = 'MacIntel'; break;
-      case 'Linux': navigatorPlatform = userAgentMetadata.architecture === 'arm' ? 'Linux aarch64' : 'Linux x86_64'; break;
-      case 'Windows': navigatorPlatform = 'Win32'; break;
-    }
-  }
-
-  return { navigatorPlatform, userAgentMetadata };
-}
-
 const paramsThatAllowContextReuse: (keyof channels.BrowserNewContextForReuseParams)[] = [
   'colorScheme',
   'forcedColors',
